refactor the common setup into abstract class
